% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{ch.bezirk}
\alias{ch.bezirk}
\title{Shapefile: Swiss district border (simplified)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 164 rows and 23 columns.
}
\usage{
ch.bezirk
}
\description{
Shapefile: Swiss district border (simplified)
}
\keyword{dataset}
