% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{ch.canton}
\alias{ch.canton}
\title{Shapefile: Swiss Canton border (simplified)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 46 rows and 22 columns.
}
\usage{
ch.canton
}
\description{
Shapefile: Swiss Canton border (simplified)
}
\keyword{dataset}
