% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{ch.ctry}
\alias{ch.ctry}
\title{Shapefile: Swiss country border (simplified)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 21 columns.
}
\usage{
ch.ctry
}
\description{
Shapefile: Swiss country border (simplified)
}
\keyword{dataset}
