% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{ch.lake}
\alias{ch.lake}
\title{Shapefile: Swiss lakes (simplified)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 55 rows and 16 columns.
}
\usage{
ch.lake
}
\description{
Shapefile: Swiss lakes (simplified)
}
\keyword{dataset}
