% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{ch.river}
\alias{ch.river}
\title{Shapefile: Swiss rivers (simplified)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13852 rows and 13 columns.
}
\usage{
ch.river
}
\description{
Shapefile: Swiss rivers (simplified)
}
\keyword{dataset}
