% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{parse_cosmo_2e}
\alias{parse_cosmo_2e}
\title{Parse COSMO-2E Point Forecasts from MeteoSwiss}
\usage{
parse_cosmo_2e(p = NULL, remote = FALSE, stn.list = NULL)
}
\arguments{
\item{p}{File path or URL}

\item{remote}{Load from remote (URL)}

\item{stn.list}{Character vector of station identifier (e. g. "SMA" or c("BAS", "ABO"))}
}
\value{
data table
}
\description{
Parse COSMO-2E Point Forecasts from MeteoSwiss
}
\examples{

require(data.table)
require(ggplot2)

# Download
dt <- parse_cosmo_2e(remote = TRUE, stn.list = c("SMA"))

# Quantiles
q <- c(.1, .9)
dt.q <- dt[, as.list(quantile(value, q, na.rm = TRUE)), .(stn, para, time)]

# Station information (including grid altitude)
attr(dt, "meta")

# Plot temperature forcast (10-90\% as ribbon, members as lines)
ggplot(dt[para == "tt"], aes(time, value)) +
  geom_ribbon(aes(y = NULL, ymin = `10\%`, ymax = `90\%`),
  data = dt.q[para == "tt"], fill = "lightblue") +
  geom_path(aes( group = member), size = .1)
}
\author{
M. Saenger
}
