% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{parse_mch_json}
\alias{parse_mch_json}
\title{Parse JSON from MeteoSwiss}
\usage{
parse_mch_json(p = NULL, para = "tt", remote = FALSE)
}
\arguments{
\item{p}{Directory path or URL}

\item{para}{Parameter}

\item{remote}{Load from remote (URL)}
}
\value{
data table
}
\description{
Source:
Description:
}
\examples{
parse_mch_json(para = "ff", remote = TRUE)
}
\author{
M. Saenger
}
