% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_mch_norm}
\alias{parse_mch_norm}
\title{Parse normal data from MeteoSwiss}
\usage{
parse_mch_norm(p = NULL, remote = FALSE, period = "1981-2010")
}
\arguments{
\item{p}{File path or URL}

\item{remote}{Load from remote (URL)}

\item{period}{Period (e. g. "1981-2010")}
}
\value{
data table
}
\description{
Parse normal data from MeteoSwiss
}
\author{
M. Saenger
}
