% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_mch_VQHA80}
\alias{parse_mch_VQHA80}
\title{Parse VQHA80.csv from MeteoSwiss}
\usage{
parse_mch_VQHA80(p = NULL, remote = FALSE)
}
\arguments{
\item{p}{File path or URL}

\item{remote}{Load from remote (URL)}
}
\value{
data table
}
\description{
Source: https://data.geo.admin.ch/ch.meteoschweiz.messwerte-aktuell/VQHA80.csv
Description: https://data.geo.admin.ch/ch.meteoschweiz.messwerte-aktuell/info/VQHA80_de.txt
}
\examples{
parse_mch_VQHA80(remote = TRUE)
}
\author{
M. Saenger
}
