% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{parse_mch_VQHA98}
\alias{parse_mch_VQHA98}
\title{Parse VQHA98.csv from MeteoSwiss (Precipitation)}
\usage{
parse_mch_VQHA98(p = NULL, remote = FALSE)
}
\arguments{
\item{p}{File path or URL}

\item{remote}{Load from remote (URL)}
}
\value{
data table
}
\description{
Source: https://data.geo.admin.ch/ch.meteoschweiz.messwerte-aktuell/VQHA98.csv
Description: https://data.geo.admin.ch/ch.meteoschweiz.messwerte-aktuell/info/VQHA98_de.txt
}
\examples{
parse_mch_VQHA98(remote = TRUE)
}
\author{
M. Saenger
}
