% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_slf_snow}
\alias{parse_slf_snow}
\title{Parse SLF snow height data}
\usage{
parse_slf_snow(p = NULL, remote = FALSE, id.var = "snow")
}
\arguments{
\item{p}{File path or URL}

\item{remote}{Load from remote (URL)}

\item{id.var}{Identifier variable (e. g. "tt" for air temperature)}
}
\value{
data table
}
\description{
Source: https://public-meas-data-v2.slf.ch/public/station-data/timepoint/SNOW_HEIGHT/current/geojson
Description: -
}
\examples{
parse_slf_snow(remote=TRUE, id.var="tt")
}
\author{
M. Saenger
}
