% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{smn.para}
\alias{smn.para}
\title{smn.para}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 105 rows and 16 columns.
}
\usage{
smn.para
}
\description{
smn.para
}
\keyword{dataset}
