% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{smn.stn.class}
\alias{smn.stn.class}
\title{SwissMetNet Station Class}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 489 rows and 3 columns.
}
\usage{
smn.stn.class
}
\description{
SwissMetNet Station Class
}
\keyword{dataset}
