% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{smn.stn.name}
\alias{smn.stn.name}
\title{SwissMetNet Station Names}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 13 rows and 2 columns.
}
\usage{
smn.stn.name
}
\description{
SwissMetNet Station Names
}
\keyword{dataset}
