% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{smn.stn.sensor.hist}
\alias{smn.stn.sensor.hist}
\title{SwissMetNet Station Sensor History}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4475 rows and 9 columns.
}
\usage{
smn.stn.sensor.hist
}
\description{
SwissMetNet Station Sensor History
}
\keyword{dataset}
