% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{smn.stn.sensor}
\alias{smn.stn.sensor}
\title{SwissMetNet Station Sensors}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2482 rows and 8 columns.
}
\usage{
smn.stn.sensor
}
\description{
SwissMetNet Station Sensors
}
\keyword{dataset}
