% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{parse_mch_norm}
\alias{parse_mch_norm}
\title{Parse normal data from MeteoSwiss}
\usage{
parse_mch_norm()
}
\arguments{
\item{p}{File path or URL}

\item{remote}{Load from remote (URL)}
}
\value{
data table
}
\description{
Parse normal data from MeteoSwiss
}
\author{
M. Saenger
}
