% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\docType{data}
\name{smn.stn}
\alias{smn.stn}
\title{SwissMetNet Stations}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 572 rows and 17 columns.
}
\usage{
smn.stn
}
\description{
SwissMetNet Stations
}
\keyword{dataset}
